extract the stringtable management to it's own class
