+++ include/osmium/output/osm/pbf.hpp
@@ -90,6 +90,9 @@ More complete outlines of real .osm.pbf files can be created using the osmpbf-ou
 // the algorithm-lib contains the sort functions
 #include <algorithm>
 
+// StringTable management
+#include <osmium/utils/stringtable.hpp>
+
 namespace Osmium {
 
     namespace Output {
@@ -219,46 +222,6 @@ namespace Osmium {
                 uint16_t primitive_block_contents;
 
                 /**
-                 * this is the struct used to build the StringTable. It is stored as
-                 * the value-part in the strings-map.
-                 *
-                 * when a new string is added to the map, its count is set to 0 and
-                 * the interim_id is set to the current size of the map. This interim_id
-                 * is then stored into the pbf-objects.
-                 *
-                 * before the PrimitiveBlock is serialized, the map is sorted by count
-                 * and stored into the pbf-StringTable. Afterwards the interim-ids are
-                 * mapped to the "real" id in the StringTable.
-                 *
-                 * this way often used strings get lower ids in the StringTable. As the
-                 * protobuf-serializer stores numbers in variable bit-lengths, lower
-                 * IDs means less used space in the resulting file.
-                 */
-                struct string_info {
-                    /**
-                     * number of occurrences of this string
-                     */
-                    uint16_t count;
-
-                    /**
-                     * an intermediate-id
-                     */
-                    uint16_t interim_id;
-                };
-
-                /**
-                 * interim StringTable, storing all strings that should be written to
-                 * the StringTable once the block is written to disk.
-                 */
-                std::map<std::string, string_info> strings;
-
-                /**
-                 * this map is used to map the interim-ids to real StringTable-IDs after
-                 * writing all strings to the StringTable.
-                 */
-                std::map<uint16_t, uint16_t> string_ids_map;
-
-                /**
                  * buffer used while compressing blobs
                  */
                 char pack_buffer[MAX_BLOB_SIZE];
@@ -281,6 +244,8 @@ namespace Osmium {
                     uint32_t user_sid;
                 } last_dense_info;
 
+                // StringTable management
+                Osmium::StringTable *st;
 
                 ///// Blob writing /////
 
@@ -395,111 +360,6 @@ namespace Osmium {
                     fwrite(data.c_str(), data.size(), 1, fd);
                 }
 
-
-                ///// StringTable management /////
-
-                /**
-                 * record a string in the interim StringTable if it's missing, otherwise just increase its counter,
-                 * return the interim-id assigned to the string.
-                 */
-                uint16_t record_string(const std::string& string) {
-                    // try to find the string in the interim StringTable
-                    if (strings.count(string) > 0) {
-                        // found, get a pointer to the associated string_info struct
-                        string_info *info_p = &strings[string];
-
-                        // increase the counter by one
-                        info_p->count++;
-
-                        // return the associated interim-id
-                        //if (Osmium::global.debug) fprintf(stderr, "found string %s at interim-id %u\n", string.c_str(), info_p->interim_id);
-                        return info_p->interim_id;
-                    } else {
-                        // not found, initialize a new string_info struct with the count set to 0 and the
-                        // interim-id set to the current size +1
-                        string_info info = {0, (uint16_t)(strings.size()+1)};
-
-                        // store this string_info struct in the interim StringTable
-                        strings[string] = info;
-
-                        // debug-print and return the associated interim-id
-                        if (Osmium::global.debug) {
-                            std::cerr << "record string " << string << " at interim-id " << info.interim_id << std::endl;
-                        }
-                        return info.interim_id;
-                    }
-                }
-
-                /**
-                 * this is the comparator used while sorting the interim StringTable.
-                 */
-                static bool stringtable_comparator(const std::pair<std::string, string_info> &a, const std::pair<std::string, string_info> &b) {
-                    // it first compares based on count
-                    if (a.second.count > b.second.count) {
-                        return true;
-                    } else if (a.second.count < b.second.count) {
-                        return false;
-                    } else {
-                        // if the count is equal, compare based on lexicography order so make
-                        // the sorting of the later zlib compression faster
-                        return a.first < b.first;
-                    }
-                }
-
-                /**
-                 * sort the interim StringTable and store it to the real protobuf StringTable.
-                 * while storing to the real table, this function fills the string_ids_map with
-                 * pairs, mapping the interim-ids to final and real StringTable ids.
-                 */
-                void store_stringtable() {
-                    // as a map can't be sorted, we need a vector holding our string/string_info pairs
-                    std::vector<std::pair<std::string, string_info>> strvec;
-
-                    // we now copy the contents from the map over to the vector
-                    std::copy(strings.begin(), strings.end(), back_inserter(strvec));
-
-                    // next the vector is sorted using our comparator
-                    std::sort(strvec.begin(), strvec.end(), stringtable_comparator);
-
-                    // add a stringtable to the PrimitiveBlock and save a pointer
-                    OSMPBF::StringTable *st = pbf_primitive_block.mutable_stringtable();
-
-                    // iterate over the items of our vector
-                    for (int i=0, l=strvec.size(); i<l; i++) {
-                        if (Osmium::global.debug) {
-                            std::cerr << "store stringtable: " << strvec[i].first << " (cnt=" << strvec[i].second.count+1 << ") with interim-id " << strvec[i].second.interim_id << " at stringtable-id " << i+1 << std::endl;
-                        }
-
-                        // add the string of the current item to the pbf StringTable
-                        st->add_s(strvec[i].first);
-
-                        // store the mapping from the interim-id to the real id
-                        string_ids_map[strvec[i].second.interim_id] = i+1;
-                    }
-                }
-
-                /**
-                 * map from an interim-id to a real string-id, throwing an exception if an
-                 * unknown mapping is requested.
-                 */
-                uint16_t map_string_id(const uint16_t interim_id) {
-                    // declare an iterator over the ids-map
-                    std::map<uint16_t, uint16_t>::const_iterator it;
-
-                    // use the find method of the map to search for the mapping pair
-                    it = string_ids_map.find(interim_id);
-
-                    // if there was a hit
-                    if (it != string_ids_map.end()) {
-                        // return the real-id stored in the second part of the pair
-                        //if (Osmium::global.debug) fprintf(stderr, "mapping interim-id %u to stringtable-id %u\n", interim_id, it->second);
-                        return it->second;
-                    }
-
-                    // throw an exception
-                    throw std::runtime_error("Request for string not in stringable\n");
-                }
-
                 /**
                  * before a PrimitiveBlock gets serialized, all interim StringTable-ids needs to be
                  * mapped to the associated real StringTable ids. Th is is done in this function.
@@ -527,7 +387,7 @@ namespace Osmium {
                                 // map interim string-ids > 0 to real string ids
                                 uint16_t sid = dense->keys_vals(i);
                                 if (sid > 0) {
-                                    dense->set_keys_vals(i, map_string_id(sid));
+                                    dense->set_keys_vals(i, st->map_string_id(sid));
                                 }
                             }
 
@@ -539,7 +399,7 @@ namespace Osmium {
                                 // iterate over all username string-ids
                                 for (int i=0, l= denseinfo->user_sid_size(); i<l; i++) {
                                     // map interim string-ids > 0 to real string ids
-                                    uint16_t user_sid = map_string_id(denseinfo->user_sid(i));
+                                    uint16_t user_sid = st->map_string_id(denseinfo->user_sid(i));
 
                                     // delta encode the string-id
                                     denseinfo->set_user_sid(i, user_sid - last_dense_info.user_sid);
@@ -572,7 +432,7 @@ namespace Osmium {
                             // iterate over all relation members, mapping the interim string-ids
                             // of the role to real string ids
                             for (int mi=0, ml=relation->roles_sid_size(); mi<ml; mi++) {
-                                relation->set_roles_sid(mi, map_string_id(relation->roles_sid(mi)));
+                                relation->set_roles_sid(mi, st->map_string_id(relation->roles_sid(mi)));
                             }
                         }
                     }
@@ -589,13 +449,13 @@ namespace Osmium {
                     if (in->has_info()) {
                         // map the interim-id of the user name to a real id
                         OSMPBF::Info *info = in->mutable_info();
-                        info->set_user_sid(map_string_id(info->user_sid()));
+                        info->set_user_sid(st->map_string_id(info->user_sid()));
                     }
 
                     // iterate over all tags and map the interim-ids of the key and the value to real ids
                     for (int i=0, l=in->keys_size(); i<l; i++) {
-                        in->set_keys(i, map_string_id(in->keys(i)));
-                        in->set_vals(i, map_string_id(in->vals(i)));
+                        in->set_keys(i, st->map_string_id(in->keys(i)));
+                        in->set_vals(i, st->map_string_id(in->vals(i)));
                     }
                 }
 
@@ -631,8 +491,8 @@ namespace Osmium {
                     // iterate over all tags and set the keys and vals, recording the strings in the
                     // interim StringTable and storing the interim ids
                     for (int i=0, l=in->tag_count(); i<l; i++) {
-                        out->add_keys(record_string(in->get_tag_key(i)));
-                        out->add_vals(record_string(in->get_tag_value(i)));
+                        out->add_keys(st->record_string(in->get_tag_key(i)));
+                        out->add_vals(st->record_string(in->get_tag_value(i)));
                     }
 
                     if (!omit_metadata()) {
@@ -642,7 +502,7 @@ namespace Osmium {
                         out_info->set_timestamp(timestamp2int(in->get_timestamp()));
                         out_info->set_changeset(in->get_changeset());
                         out_info->set_uid(in->get_uid());
-                        out_info->set_user_sid(record_string(in->get_user()));
+                        out_info->set_user_sid(st->record_string(in->get_user()));
                     }
                 }
 
@@ -671,7 +531,7 @@ namespace Osmium {
                     }
 
                     // store the interim StringTable into the protobuf object
-                    store_stringtable();
+                    st->store_stringtable(pbf_primitive_block.mutable_stringtable());
 
                     // map all interim string ids to real ids
                     map_string_ids();
@@ -691,9 +551,8 @@ namespace Osmium {
                     pbf_primitive_block.set_granularity(granularity());
                     pbf_primitive_block.set_date_granularity(date_granularity());
 
-                    // clear the interim StringTable and its id map
-                    strings.clear();
-                    string_ids_map.clear();
+                    // clear the StringTable
+                    st->clear();
 
                     // reset the dense-info struct to zero
                     last_dense_info = {0, 0, 0, 0, 0, 0, 0};
@@ -767,8 +626,8 @@ namespace Osmium {
                     // the first node has two tags (3=>5 and 2=>1), the second node has does not
                     // have any tags and the third node has a single tag (8=>5)
                     for (int i=0, l=node->tag_count(); i<l; i++) {
-                        dense->add_keys_vals(record_string(node->get_tag_key(i)));
-                        dense->add_keys_vals(record_string(node->get_tag_value(i)));
+                        dense->add_keys_vals(st->record_string(node->get_tag_key(i)));
+                        dense->add_keys_vals(st->record_string(node->get_tag_value(i)));
                     }
                     dense->add_keys_vals(0);
 
@@ -796,7 +655,7 @@ namespace Osmium {
 
                         // record the user-name to the interim stringtable and copy the
                         // interim string-id to the pbf-object
-                        denseinfo->add_user_sid(record_string(node->get_user()));
+                        denseinfo->add_user_sid(st->record_string(node->get_user()));
                     }
                 }
 
@@ -842,7 +701,7 @@ namespace Osmium {
 
                         // record the relation-member role to the interim stringtable and copy the
                         // interim string-id to the pbf-object
-                        pbf_relation->add_roles_sid(record_string(mem->get_role()));
+                        pbf_relation->add_roles_sid(st->record_string(mem->get_role()));
 
                         // copy the relation-member-id, delta encoded against last_id
                         int64_t id = mem->get_ref();
@@ -881,6 +740,9 @@ namespace Osmium {
                     // set defaults from proto-definiton
                     granularity_ = pbf_primitive_block.granularity();
                     date_granularity_ = pbf_primitive_block.date_granularity();
+
+                    // create StringTable management object
+                    st = new Osmium::StringTable();
                 }
 
                 /**
@@ -903,11 +765,18 @@ namespace Osmium {
                     granularity_ = pbf_primitive_block.granularity();
                     date_granularity_ = pbf_primitive_block.date_granularity();
 
+                    // create StringTable management object
+                    st = new Osmium::StringTable();
+
                     // open the file for writing
                     fd = fopen(filename.c_str(), "w");
                     if (!fd) {
                         throw std::runtime_error("unable to open outfile");
                     }
+                }
+
+                ~PBF() {
+                    delete st;
                 }
